[Silver II] Title: 이진수 격자, Time: 68 ms, Memory: 11832 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/31910. 이진수 격자/이진수 격자.java
@@ -13,7 +13,7 @@ public class Main {
 
         n = Integer.parseInt(br.readLine());
         int[][] arr = new int[n+1][n+1];
-        int[][] dp = new int[n+1][n+1];
+        long[][] dp = new long[n+1][n+1];
         for(int i=1;i<=n;i++){
             StringTokenizer st = new StringTokenizer(br.readLine());
             for(int j=1;j<=n;j++){
